Update README.md to make examples more clear
Most people seem to be confused by the example, where `'token'` is the first element of the `registration_ids` array, and three dots the second element. https://github.com/rpush/rpush/issues/96

For consistency with the iOS example, where the `device_token` is displayed as `'...'`, and to avoid the confusion described in the linked issue, I propose this change.

Thanks for such a great library. It's been a lifesaver.