Update jruby to 1.7.23, allow failures for jruby
I know this is not fixing the problem, but I don't know enough at this
point to fix this issue. I found out that the `multi_json` gem is not
getting installed for jRuby. After explicitly setting it as a dependency
in the Gemfile, the next gem was missing, namely `ansi`.

To me, it appears that the runtime dependencies declared in the
`rpush.gemspec` file are ignored for jRuby build. I don't know why. :\