Fail gracefully when a Modis notification no longer exists
This makes behavior more consistent with the ActiveRecord adapter. If a service attempts to call #mark_ids_retryable or #mark_ids_failed and includes an ID that no longer exists in the datastore, the adapter will simply skip that record and continue, rather than crashing. (Currently, the adapter crashes, which can cause delivery issues with other unrelated notifications.)